chore: seperate the server and listener
